requirements.txt
SQLAlchemy~=2.0.38
python-dotenv~=1.1.1
uvicorn~=0.35.0
fastapi~=0.116.1
PyJWT~=2.10.1
pydantic~=2.11.7
argon2-cffi~=23.1.0
email-validator~=2.3.0
